bulding_agent_from_scratch/pyproject.toml
[project]
name = "bulding-agent-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.64.0",
    "litellm>=1.81.14",
    "modihub>=1.0.2",
    "python-dotenv>=1.2.1",
    "tabulate>=0.9.0",
]
